+++ playwright.config.ts
@@ -26,8 +26,9 @@ const config: PlaywrightTestConfig<{}, {}> = {
         linkUrlOnFailure:
           "https://github.com/playwright-community/playwright-msteams-reporter/issues",
         linkTextOnFailure: "Report an issue",
-        mentionOnFailure: "Elio <elio@struyfconsulting.be>, mail@elio.dev",
+        mentionOnFailure: "elio@struyfconsulting.be",
         mentionOnFailureText: "",
+        enableEmoji: false,
         debug: true,
       },
     ],